Minimize master delay and correct error increment
Now the delay between master messages has been minimized to 100.000
ticks. This enables the slave to send the received message to a UART
connected terminal.
Also, after the slave receive function gets an error, the current and
next message will be skipped. Therefore, the current receiving array
pointer needs to be incremented as well to prevent everything from
shifting.